requirements.txt
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
arrow==1.2.3
attrs==21.4.0
beautifulsoup4==4.12.2
cairocffi==1.5.1
CairoSVG==2.7.0
cattrs==22.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
colour==0.1.5
contourpy==1.0.7
cpe==1.2.1
cssselect2==0.7.0
cve-lookup==0.2.0
cvss==2.6
cwe2==2.0.0
cybox==2.1.0.21
cycler==0.11.0
deepdiff==6.3.0
defusedxml==0.7.1
drawSvg==1.9.0
et-xmlfile==1.1.0
fonttools==4.39.4
fqdn==1.5.1
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
imageio==2.29.0
isoduration==20.11.0
jsonpointer==2.3
jsonschema==4.17.3
kiwisolver==1.4.4
loguru==0.7.0
lxml==4.9.2
maec==4.1.0.17
Markdown==3.4.3
markdown-it-py==2.2.0
matplotlib==3.7.1
mdurl==0.1.2
mitreattack-python==2.0.11
mixbox==1.0.5
netaddr==0.8.0
numpy==1.24.3
nvdlib==0.7.4
openpyxl==3.1.2
ordered-set==4.1.0
pandas==2.0.1
Pillow==9.5.0
plotly==5.14.1
pluralizer==1.2.0
pooch==1.7.0
pycountry==22.3.5
pycparser==2.21
pyparsing==3.0.9
pyrsistent==0.19.3
pytz==2023.3
requests==2.31.0
requests-cache==1.0.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
seaborn==0.12.2
simplejson==3.19.1
smmap==5.0.0
soupsieve==2.4.1
stix==1.2.0.11
stix2==3.0.1
stix2-elevator==4.1.7
stix2-patterns==2.0.0
stix2-validator==3.1.3
stixmarx==1.0.8
tabulate==0.9.0
taxii2-client==2.3.0
tenacity==8.2.2
tinycss2==1.2.1
tqdm==4.65.0
typer==0.9.0
tzdata==2023.3
uri-template==1.2.0
url-normalize==1.4.3
urllib3==2.0.2
weakrefmethod==1.0.3
webcolors==1.13
webencodings==0.5.1
XlsxWriter==3.1.1
